Parsen toegevoegd op de firstusePage en db aan models toegevoegd.

+++ Time2WorkApp/Time2WorkApp.Android/MainActivity.cs
@@ -124,7 +124,7 @@ namespace Time2WorkApp.Droid
         public int id { get; set; }
         public string firstname { get; set; }
         public string lastname { get; set; }
-        public int brutoloon { get; set; }
+        public double brutoloon { get; set; }
         public string email { get; set; }
         public string password { get; set; }
     }
